requirements-api.txt
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
pandas>=2.0,<3.0
pycountry>=22.0,<24.0
pyyaml>=6.0,<7.0